third_party/blink/web_tests/external/wpt/editing/run/multitest_1-1000-expected.txt

This is a testharness.js-based test.
Found 205 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["bold",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<b>a</b>bar" but got "foo<span style=\\"font-weight:bold\\">a</span>bar"
[FAIL] [["bold",""],["inserttext","a"]] "foo[]bar" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["inserttext","a"]] "foo[]bar" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""],["delete",""]] "foo[]bar" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["delete",""]] "foo[]bar" queryCommandState("bold") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["bold",""],["delete",""]] "foo[]bar" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["delete",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "fo<b>a</b>bar" but got "foabar"
[FAIL] [["bold",""],["delete",""],["inserttext","a"]] "foo[]bar" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["delete",""],["inserttext","a"]] "foo[]bar" queryCommandState("bold") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["bold",""],["delete",""],["inserttext","a"]] "foo[]bar" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["formatblock","<div>"]] "foo[]bar" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["formatblock","<div>"]] "foo[]bar" queryCommandState("bold") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["bold",""],["formatblock","<div>"]] "foo[]bar" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["formatblock","<div>"],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo<b>a</b>bar</div>" but got "<div>fooabar</div>"
[FAIL] [["bold",""],["formatblock","<div>"],["inserttext","a"]] "foo[]bar" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["formatblock","<div>"],["inserttext","a"]] "foo[]bar" queryCommandState("bold") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["bold",""],["formatblock","<div>"],["inserttext","a"]] "foo[]bar" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["forwarddelete",""]] "foo[]bar" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["forwarddelete",""]] "foo[]bar" queryCommandState("bold") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["bold",""],["forwarddelete",""]] "foo[]bar" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["forwarddelete",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<b>a</b>ar" but got "fooaar"
[FAIL] [["bold",""],["forwarddelete",""],["inserttext","a"]] "foo[]bar" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["forwarddelete",""],["inserttext","a"]] "foo[]bar" queryCommandState("bold") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["bold",""],["forwarddelete",""],["inserttext","a"]] "foo[]bar" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["indent",""]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote>foobar</blockquote>" but got "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">foobar</blockquote>"
[FAIL] [["bold",""],["indent",""]] "foo[]bar" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["indent",""]] "foo[]bar" queryCommandState("bold") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["bold",""],["indent",""]] "foo[]bar" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["indent",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote>foo<b>a</b>bar</blockquote>" but got "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">fooabar</blockquote>"
[FAIL] [["bold",""],["indent",""],["inserttext","a"]] "foo[]bar" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["indent",""],["inserttext","a"]] "foo[]bar" queryCommandState("bold") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["bold",""],["indent",""],["inserttext","a"]] "foo[]bar" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["inserthorizontalrule",""]] "foo[]bar" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["inserthorizontalrule",""]] "foo[]bar" queryCommandState("bold") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["bold",""],["inserthorizontalrule",""]] "foo[]bar" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["inserthorizontalrule",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<hr><b>a</b>bar" but got "foo<hr>abar"
[FAIL] [["bold",""],["inserthorizontalrule",""],["inserttext","a"]] "foo[]bar" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["inserthorizontalrule",""],["inserttext","a"]] "foo[]bar" queryCommandState("bold") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["bold",""],["inserthorizontalrule",""],["inserttext","a"]] "foo[]bar" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["inserthtml","ab<b>c</b>d"]] "foo[]bar" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["inserthtml","ab<b>c</b>d"]] "foo[]bar" queryCommandState("bold") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["bold",""],["inserthtml","ab<b>c</b>d"]] "foo[]bar" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["inserthtml","ab<b>c</b>d"],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "fooab<b>c</b>d<b>a</b>bar" but got "fooab<b>c</b>dabar"
[FAIL] [["bold",""],["inserthtml","ab<b>c</b>d"],["inserttext","a"]] "foo[]bar" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["inserthtml","ab<b>c</b>d"],["inserttext","a"]] "foo[]bar" queryCommandState("bold") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["bold",""],["inserthtml","ab<b>c</b>d"],["inserttext","a"]] "foo[]bar" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["insertimage","/img/lion.svg"]] "foo[]bar" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["insertimage","/img/lion.svg"]] "foo[]bar" queryCommandState("bold") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["bold",""],["insertimage","/img/lion.svg"]] "foo[]bar" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["insertimage","/img/lion.svg"],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<img src=\\"/img/lion.svg\\"><b>a</b>bar" but got "foo<img src=\\"/img/lion.svg\\">abar"
[FAIL] [["bold",""],["insertimage","/img/lion.svg"],["inserttext","a"]] "foo[]bar" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["insertimage","/img/lion.svg"],["inserttext","a"]] "foo[]bar" queryCommandState("bold") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["bold",""],["insertimage","/img/lion.svg"],["inserttext","a"]] "foo[]bar" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["insertlinebreak",""]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<br>bar" but got "foo<span style=\\"font-weight:bold\\"><br></span>bar"
[FAIL] [["bold",""],["insertlinebreak",""]] "foo[]bar" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["insertlinebreak",""]] "foo[]bar" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""],["insertlinebreak",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<br><b>a</b>bar" but got "foo<span style=\\"font-weight:bold\\"><br>a</span>bar"
[FAIL] [["bold",""],["insertlinebreak",""],["inserttext","a"]] "foo[]bar" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["insertlinebreak",""],["inserttext","a"]] "foo[]bar" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""],["insertorderedlist",""]] "foo[]bar" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["insertorderedlist",""]] "foo[]bar" queryCommandState("bold") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["bold",""],["insertorderedlist",""]] "foo[]bar" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["insertorderedlist",""]] "foo[]bar" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["insertorderedlist",""]] "foo[]bar" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo<b>a</b>bar</li></ol>" but got "<ol><li>afoobar</li></ol>"
[FAIL] [["bold",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandState("bold") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["bold",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""],["insertparagraph",""]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>bar</p>" but got "foo<div>bar</div>"
[FAIL] [["bold",""],["insertparagraph",""]] "foo[]bar" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["insertparagraph",""]] "foo[]bar" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""],["insertparagraph",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p><b>a</b>bar</p>" but got "foo<div><span style=\\"font-weight:bold\\">a</span>bar</div>"
[FAIL] [["bold",""],["insertparagraph",""],["inserttext","a"]] "foo[]bar" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["insertparagraph",""],["inserttext","a"]] "foo[]bar" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""],["insertunorderedlist",""]] "foo[]bar" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["insertunorderedlist",""]] "foo[]bar" queryCommandState("bold") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["bold",""],["insertunorderedlist",""]] "foo[]bar" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["insertunorderedlist",""]] "foo[]bar" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["insertunorderedlist",""]] "foo[]bar" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo<b>a</b>bar</li></ul>" but got "<ul><li>afoobar</li></ul>"
[FAIL] [["bold",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandState("bold") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["bold",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""],["justifycenter",""]] "foo[]bar" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["justifycenter",""]] "foo[]bar" queryCommandState("bold") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["bold",""],["justifycenter",""]] "foo[]bar" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["justifycenter",""]] "foo[]bar" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["bold",""],["justifycenter",""]] "foo[]bar" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["bold",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\">foo<b>a</b>bar</div>" but got "<div style=\\"text-align:center\\">fooabar</div>"
[FAIL] [["bold",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandState("bold") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["bold",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["bold",""],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["bold",""],["justifyfull",""]] "foo[]bar" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["justifyfull",""]] "foo[]bar" queryCommandState("bold") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["bold",""],["justifyfull",""]] "foo[]bar" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["justifyfull",""]] "foo[]bar" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["bold",""],["justifyfull",""]] "foo[]bar" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["bold",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\">foo<b>a</b>bar</div>" but got "<div style=\\"text-align:justify\\">fooabar</div>"
[FAIL] [["bold",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandState("bold") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["bold",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["bold",""],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["bold",""],["justifyleft",""]] "foo[]bar" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["justifyleft",""]] "foo[]bar" queryCommandState("bold") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["bold",""],["justifyleft",""]] "foo[]bar" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["justifyleft",""]] "foo[]bar" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["bold",""],["justifyleft",""]] "foo[]bar" queryCommandValue("justifyleft") after
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["bold",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<b>a</b>bar" but got "fooabar"
[FAIL] [["bold",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandState("bold") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["bold",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["bold",""],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyleft") after
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["bold",""],["justifyright",""]] "foo[]bar" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["justifyright",""]] "foo[]bar" queryCommandState("bold") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["bold",""],["justifyright",""]] "foo[]bar" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["justifyright",""]] "foo[]bar" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["bold",""],["justifyright",""]] "foo[]bar" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["bold",""],["justifyright",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\">foo<b>a</b>bar</div>" but got "<div style=\\"text-align:right\\">fooabar</div>"
[FAIL] [["bold",""],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandState("bold") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["bold",""],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["bold",""],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["bold",""],["outdent",""]] "foo[]bar" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["outdent",""]] "foo[]bar" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["bold",""],["outdent",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<b>a</b>bar" but got "foo<span style=\\"font-weight:bold\\">a</span>bar"
[FAIL] [["bold",""],["outdent",""],["inserttext","a"]] "foo[]bar" queryCommandValue("bold") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["bold",""],["outdent",""],["inserttext","a"]] "foo[]bar" queryCommandValue("bold") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["italic",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<i>a</i>bar" but got "foo<span style=\\"font-style:italic\\">a</span>bar"
[FAIL] [["italic",""],["inserttext","a"]] "foo[]bar" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["inserttext","a"]] "foo[]bar" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["italic",""],["delete",""]] "foo[]bar" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["delete",""]] "foo[]bar" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["italic",""],["delete",""]] "foo[]bar" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["delete",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "fo<i>a</i>bar" but got "foabar"
[FAIL] [["italic",""],["delete",""],["inserttext","a"]] "foo[]bar" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["delete",""],["inserttext","a"]] "foo[]bar" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["italic",""],["delete",""],["inserttext","a"]] "foo[]bar" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["formatblock","<div>"]] "foo[]bar" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["formatblock","<div>"]] "foo[]bar" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["italic",""],["formatblock","<div>"]] "foo[]bar" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["formatblock","<div>"],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo<i>a</i>bar</div>" but got "<div>fooabar</div>"
[FAIL] [["italic",""],["formatblock","<div>"],["inserttext","a"]] "foo[]bar" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["formatblock","<div>"],["inserttext","a"]] "foo[]bar" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["italic",""],["formatblock","<div>"],["inserttext","a"]] "foo[]bar" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["forwarddelete",""]] "foo[]bar" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["forwarddelete",""]] "foo[]bar" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["italic",""],["forwarddelete",""]] "foo[]bar" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["forwarddelete",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<i>a</i>ar" but got "fooaar"
[FAIL] [["italic",""],["forwarddelete",""],["inserttext","a"]] "foo[]bar" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["forwarddelete",""],["inserttext","a"]] "foo[]bar" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["italic",""],["forwarddelete",""],["inserttext","a"]] "foo[]bar" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["indent",""]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote>foobar</blockquote>" but got "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">foobar</blockquote>"
[FAIL] [["italic",""],["indent",""]] "foo[]bar" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["indent",""]] "foo[]bar" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["italic",""],["indent",""]] "foo[]bar" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["indent",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote>foo<i>a</i>bar</blockquote>" but got "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">fooabar</blockquote>"
[FAIL] [["italic",""],["indent",""],["inserttext","a"]] "foo[]bar" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["indent",""],["inserttext","a"]] "foo[]bar" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["italic",""],["indent",""],["inserttext","a"]] "foo[]bar" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["inserthorizontalrule",""]] "foo[]bar" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["inserthorizontalrule",""]] "foo[]bar" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["italic",""],["inserthorizontalrule",""]] "foo[]bar" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["inserthorizontalrule",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<hr><i>a</i>bar" but got "foo<hr>abar"
[FAIL] [["italic",""],["inserthorizontalrule",""],["inserttext","a"]] "foo[]bar" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["inserthorizontalrule",""],["inserttext","a"]] "foo[]bar" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["italic",""],["inserthorizontalrule",""],["inserttext","a"]] "foo[]bar" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["inserthtml","ab<b>c</b>d"]] "foo[]bar" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["inserthtml","ab<b>c</b>d"]] "foo[]bar" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["italic",""],["inserthtml","ab<b>c</b>d"]] "foo[]bar" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["inserthtml","ab<b>c</b>d"],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "fooab<b>c</b>d<i>a</i>bar" but got "fooab<b>c</b>dabar"
[FAIL] [["italic",""],["inserthtml","ab<b>c</b>d"],["inserttext","a"]] "foo[]bar" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["inserthtml","ab<b>c</b>d"],["inserttext","a"]] "foo[]bar" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["italic",""],["inserthtml","ab<b>c</b>d"],["inserttext","a"]] "foo[]bar" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["insertimage","/img/lion.svg"]] "foo[]bar" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["insertimage","/img/lion.svg"]] "foo[]bar" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["italic",""],["insertimage","/img/lion.svg"]] "foo[]bar" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["insertimage","/img/lion.svg"],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<img src=\\"/img/lion.svg\\"><i>a</i>bar" but got "foo<img src=\\"/img/lion.svg\\">abar"
[FAIL] [["italic",""],["insertimage","/img/lion.svg"],["inserttext","a"]] "foo[]bar" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["insertimage","/img/lion.svg"],["inserttext","a"]] "foo[]bar" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["italic",""],["insertimage","/img/lion.svg"],["inserttext","a"]] "foo[]bar" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["insertlinebreak",""]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<br>bar" but got "foo<span style=\\"font-style:italic\\"><br></span>bar"
[FAIL] [["italic",""],["insertlinebreak",""]] "foo[]bar" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["insertlinebreak",""]] "foo[]bar" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["italic",""],["insertlinebreak",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<br><i>a</i>bar" but got "foo<span style=\\"font-style:italic\\"><br>a</span>bar"
[FAIL] [["italic",""],["insertlinebreak",""],["inserttext","a"]] "foo[]bar" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["insertlinebreak",""],["inserttext","a"]] "foo[]bar" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["italic",""],["insertorderedlist",""]] "foo[]bar" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["insertorderedlist",""]] "foo[]bar" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["italic",""],["insertorderedlist",""]] "foo[]bar" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["insertorderedlist",""]] "foo[]bar" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["insertorderedlist",""]] "foo[]bar" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["italic",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo<i>a</i>bar</li></ol>" but got "<ol><li>afoobar</li></ol>"
[FAIL] [["italic",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["italic",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
Harness: the test ran to completion.
